chore: Migrate CreateSessionPresenter to ViewModel
