+++ app/src/main/java/com/eventyay/organizer/core/session/create/CreateSessionFragment.java
@@ -3,6 +3,9 @@ package com.eventyay.organizer.core.session.create;
 import androidx.databinding.DataBindingUtil;
 import android.os.Bundle;
 import androidx.annotation.Nullable;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.lifecycle.ViewModelProviders;
+
 import com.google.android.material.textfield.TextInputLayout;
 import android.text.Editable;
 import android.text.TextUtils;
@@ -25,20 +28,21 @@ import com.eventyay.organizer.utils.ValidateUtils;
 import javax.inject.Inject;
 
 import br.com.ilhasoft.support.validation.Validator;
-import dagger.Lazy;
 
 import static com.eventyay.organizer.ui.ViewUtils.showView;
 
-public class CreateSessionFragment extends BaseFragment<CreateSessionPresenter> implements CreateSessionView {
+public class CreateSessionFragment extends BaseFragment implements CreateSessionView {
 
     @Inject
-    Lazy<CreateSessionPresenter> presenterProvider;
+    ViewModelProvider.Factory viewModelFactory;
 
     private SessionCreateLayoutBinding binding;
     private Validator validator;
     public static final String TRACK_KEY = "track";
     private static final String SESSION_KEY = "session_id";
     private ArrayAdapter<CharSequence> sessionStateAdapter;
+
+    private CreateSessionViewModel createSessionViewModel;
 
     private boolean isSessionUpdating;
     private long trackId;
@@ -84,14 +88,15 @@ public class CreateSessionFragment extends BaseFragment<CreateSessionPresenter> 
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         binding = DataBindingUtil.inflate(inflater, R.layout.session_create_layout, container, false);
+        createSessionViewModel = ViewModelProviders.of(this, viewModelFactory).get(CreateSessionViewModel.class);
         validator = new Validator(binding.form);
 
         binding.submit.setOnClickListener(view -> {
             if (validator.validate()) {
                 if (isSessionUpdating) {
-                    getPresenter().updateSession(trackId, eventId);
+                    createSessionViewModel.updateSession(trackId, eventId);
                 } else {
-                    getPresenter().createSession(trackId, eventId);
+                    createSessionViewModel.createSession(trackId, eventId);
                 }
             }
         });
@@ -110,11 +115,15 @@ public class CreateSessionFragment extends BaseFragment<CreateSessionPresenter> 
     @Override
     public void onStart() {
         super.onStart();
-        getPresenter().attach(this);
-        binding.setSession(getPresenter().getSession());
+        createSessionViewModel.getProgress().observe(this, this::showProgress);
+        createSessionViewModel.getDismiss().observe(this, (dismiss) -> dismiss());
+        createSessionViewModel.getSuccess().observe(this, this::onSuccess);
+        createSessionViewModel.getError().observe(this, this::showError);
+        createSessionViewModel.getSessionLiveData().observe(this, this::setSession);
+        binding.setSession(createSessionViewModel.getSession());
 
         if (isSessionUpdating) {
-            getPresenter().loadSession(sessionId);
+            createSessionViewModel.loadSession(sessionId);
         }
 
         validate(binding.form.slidesUrlLayout, ValidateUtils::validateUrl, getResources().getString(R.string.url_validation_error));
@@ -157,7 +166,7 @@ public class CreateSessionFragment extends BaseFragment<CreateSessionPresenter> 
     @Override
     public void setSession(Session session) {
         binding.setSession(session);
-        String state = getPresenter().getSession().getState();
+        String state = createSessionViewModel.getSession().getState();
         int statePosition = sessionStateAdapter.getPosition(state);
         binding.form.spinner.setSelection(statePosition);
     }
@@ -169,11 +178,6 @@ public class CreateSessionFragment extends BaseFragment<CreateSessionPresenter> 
         } else {
             return R.string.create_session;
         }
-    }
-
-    @Override
-    public Lazy<CreateSessionPresenter> getPresenterProvider() {
-        return presenterProvider;
     }
 
     @Override
